Refactor statistics sections to use generic data
The StatisticsGeneralSection and StatisticsPlayerSection components were refactored to use a
generic data object instead of specific types. This allows for more flexibility in the data
that can be displayed in the statistics sections.

The formatMs function was renamed to formatMsToHoursMins and a new formatMs function was
created to format milliseconds into hours and minutes.